+++ src/classic/avdtp_source.c
@@ -176,7 +176,7 @@ void avdtp_source_init(avdtp_context_t * avdtp_context){
     avdtp_source_context->stream_endpoints_id_counter = 0;
     avdtp_source_context->packet_handler = packet_handler;
 
-    l2cap_register_service(&packet_handler, BLUETOOTH_PSM_AVDTP, 0xffff, LEVEL_2);
+    l2cap_register_service(&packet_handler, BLUETOOTH_PSM_AVDTP, 0xffff, gap_get_security_level());
 }