+++ src/hci_transport_h5.c
@@ -741,7 +741,7 @@ static void hci_transport_h5_read_next_byte(void){
 
 // track time receiving SLIP frame
 static uint32_t hci_transport_h5_receive_start;
-static void hci_transport_h5_block_received(){
+static void hci_transport_h5_block_received(void){
     if (hci_transport_h5_active == 0) return;
 
     // track start time when receiving first byte // a bit hackish
@@ -753,8 +753,10 @@ static void hci_transport_h5_block_received(){
     if (frame_size) {
         // track time
         uint32_t packet_receive_time = btstack_run_loop_get_time_ms() - hci_transport_h5_receive_start;
-        uint32_t nominmal_time = (frame_size + 6) * 10 * 1000 / uart_config.baudrate;
-        log_info("slip frame time %u ms for %u decoded bytes. nomimal time %u ms", (int) packet_receive_time, frame_size, (int) nominmal_time);
+        uint32_t nominal_time = (frame_size + 6) * 10 * 1000 / uart_config.baudrate;
+        UNUSED(nominal_time);
+        UNUSED(packet_receive_time);
+        log_info("slip frame time %u ms for %u decoded bytes. nomimal time %u ms", (int) packet_receive_time, frame_size, (int) nominal_time);
         // reset state
         hci_transport_h5_receive_start = 0;
         //